add crontab internal

+++ application/controllers/internal/Missing.php
@@ -25,7 +25,7 @@ class Missing extends MY_Controller {
             $a_set_reward = $this->campaign_model->get_set_reward($a_conversion_missing['campaign_id']);
     
             $reward = !empty($a_set_reward) ? $a_set_reward['existing'] : 0.1;
-            $summary_reward = (int)$a_conversion_missing['amount'] * ((int)$reward/100);
+            $summary_reward = (float)$a_conversion_missing['amount'] * ((float)$reward/100);
             
             $this->report_conversion_model->update_by_conversion_id2(
                 $a_conversion_missing['id'],
